updata syokisyoki updata()

+++ Assets/ParameterData.cs
@@ -6,10 +6,10 @@ public class ParameterData
 {
     public int Score { get; private set; }
     public int BlackDepth { get; private set; }
-    public double Speed { get; private set; }
+    public float Speed { get; private set; }
     public int Credit { get; private set; }
     
-    public ParameterData(int Score, int BlackDepth, double Speed, int Credit)
+    public ParameterData(int Score, int BlackDepth, float Speed, int Credit)
     {
         this.Score = Score;
         this.BlackDepth = BlackDepth;
